requirements.txt
# GPU Render Server Requirements

# Core frameworks
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Web automation & rendering
playwright==1.40.0
selenium==4.16.0

# Image/Video processing
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3

# AWS services
boto3==1.29.7
botocore==1.32.7

# Redis for job queue
redis==5.0.1

# HTTP client
aiohttp==3.9.1
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# GPU support (optional, install separately)
# torch==2.0.1+cu118
# torchvision==0.15.2+cu118